pyproject.toml
[project]
name = "marklogic"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "opencv-python>=4.13.0.92",
    "pillow>=12.1.1",
    "tqdm>=4.67.3",
    "ultralytics>=8.4.21",
]
